# src/Dfe.CaseAggregationService.Application/Schools/Models/NameDetailsModel.cs
namespace Dfe.CaseAggregationService.Application.Schools.Models
{
    public class NameDetailsModel
    {
        public string? FirstName { get; set; }
        public string? LastName { get; set; }
        public string? MiddleName { get; set; }
    }
}